feature/JWT-and-fixes

See merge request apostolos-javascript/assina-club!18

+++ src/pages/api/subscribers/[subscriberId].ts
@@ -1,5 +1,7 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
 
+import bcrypt from 'bcrypt'
+
 import { SubscriberType, Subscriber } from '../../../@types/SubscriberTypes'
 
 import { getSubscriber } from '../../../prisma/subscribers'
@@ -33,6 +35,12 @@ export default async function updateSubscriber(
             unsubscribe
         }: Subscriber = req.body
 
+        const subscriberExists = await getSubscriber(subscriberId)
+
+        if(!subscriberExists) return res.status(404).json({
+            message: 'Subscriber does not exist'
+        })
+
         if (unsubscribe) {
             if (!clubAssinatureId) {
                 return res.status(400).json({
@@ -61,6 +69,8 @@ export default async function updateSubscriber(
             })
         }
 
+        const hashedPassword = bcrypt.hashSync(password, 10)
+
         const subscriber = await prisma.subscriber.update({
             where: { id: subscriberId },
             data: {
@@ -68,7 +78,7 @@ export default async function updateSubscriber(
                 cpf,
                 birthDate,
                 email,
-                password
+                password: hashedPassword
             }
         })
 
@@ -78,6 +88,13 @@ export default async function updateSubscriber(
         })
 
     } else if (method === "DELETE") {
+
+        const subscriberExists = await getSubscriber(subscriberId)
+
+        if(!subscriberExists) return res.status(404).json({
+            message: 'Subscriber does not exist'
+        })
+
         await prisma.subscriber.delete({
             where: { id: subscriberId }
         })
